+++ database-lib/src/main/java/com/mba/logic/database_lib/coroutine/IHDLRepos.kt
@@ -5,7 +5,7 @@ package com.mba.logic.database_lib.coroutine
  */
 interface IHDLRepos {
     fun <T> query(func: () -> T, err: ((e: Exception) -> Unit)? = null, callback: (result: T) -> Unit)
-    fun insert(func: () -> Int, err: ((e: Exception) -> Unit)? = null, callback: (result: Int) -> Unit)
+    fun insert(func: () -> Long, err: ((e: Exception) -> Unit)? = null, callback: (result: Long) -> Unit)
     fun update(func: () -> Int, err: ((e: Exception) -> Unit)? = null, callback: (result: Int) -> Unit)
     fun delete(func: () -> Int, err: ((e: Exception) -> Unit)? = null, callback: (result: Int) -> Unit)
     fun transaction(vararg func: () -> Unit, err: ((e: Exception) -> Unit)? = null, callback: () -> Unit)
